+++ src/atf_model.ts
@@ -1,4 +1,4 @@
-import { EOL } from 'os';
+import * as nisabaLogger from './logger'
 
 /**
  * Extract the project code from an ATF file.
@@ -9,10 +9,14 @@ import { EOL } from 'os';
  */
 export function getProjectCode(atfText: string): string {
     const tag = "#project:";
-    const codes = atfText.split(EOL)
+    // We can't use `os.EOL` to split the lines because that'd be OS-specific,
+    // that's the point of that variable, but we need to handle also CRLF files
+    // on Unix systems and LF files on Windows.
+    const codes = atfText.split(/\r?\n/)
                     .filter(line => line.startsWith(tag))
                     .map(line => line.slice(tag.length).trim());
     if ((new Set(codes)).size == 1) {
+        nisabaLogger.debug(`Project code found: "${codes[0]}"`);
         return codes[0];
     }
     if (codes.length == 0) {